feat(predict-track-cart): Emarsys trackCart delegates to predictInternal and shardRepository SUITEDEV-13717

+++ predict-api/src/main/java/com/emarsys/predict/api/model/CartItem.java
@@ -2,6 +2,6 @@ package com.emarsys.predict.api.model;
 
 public interface CartItem {
     String getItemId();
-    int getPrice();
-    int getQuantity();
+    double getPrice();
+    double getQuantity();
 }
